Add network migrations
Add capabilites to migrate pages directly from device to device via a 100G Ethernet connection.
The newly added NetworkPageDMA core extends the existing DMA core with a AXI4 Stream interface to transmit and receive data.
The EthernetFlowControl core triggers pause frames to sender side if the receiver side is temporarily congested.

On host side the driver checks for device-to-device migrations whether both devices are capable of network migrations and uses the method then.

TODO: Use proper MAC addresses. These must be set in the DMA core but also in the 100G core for pause frames.