Track migrated regions in interval tree
Switch to an interval tree to track which virtual memory intervals are migrated to a specific Tapasco device. We use the standard interval tree implementation in the kernel.
Since we do not track the corresponding physical addresses in the interval tree to save space, we re-introduce the PTW to perform a virtual to physical address translation.